+++ R/zzz.R
@@ -7,7 +7,6 @@
   s3_register("dplyr::summarise", "cubble_df")
   s3_register("dplyr::ungroup", "cubble_df")
   s3_register("dplyr::group_by", "cubble_df")
-  s3_register("dplyr::select", "cubble_df")
 
   s3_register("dplyr::slice_head", "cubble_df")
   s3_register("dplyr::slice_tail", "cubble_df")
@@ -16,5 +15,6 @@
   s3_register("dplyr::slice_sample", "cubble_df")
 
   s3_register("dplyr::rename", "cubble_df")
-
+  s3_register("base::print", "cubble_df")
+  s3_register("tsibble::fill_gaps", "cubble_df")
 }