+++ NEWS.md
@@ -1,3 +1,9 @@
+# cubble 0.2.0
+
+* update on `geom_glyph` to take regular function
+* a new vignette on *making a linked map and time series plot*
+* add a converter between `stars` and `cubble` object #5
+
 # cubble 0.1.1
 
 * remove dependency (suggest) on `rnoaa`